requirements.txt
aioboto3==7.0.0
aiobotocore==2.17.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aioitertools==0.12.0
aiosignal==1.4.0
aiosonic==0.24.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
asyncio==3.4.3
attrs==25.3.0
beautifulsoup4==4.13.4
bitarray==3.5.0
blinker==1.9.0
boto3==1.35.90
botocore==1.35.93
cachetools==6.1.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
ckzg==2.1.1
click==8.2.1
cmake==4.0.3
coincurve==21.0.0
colorama==0.4.6
contourpy==1.3.2
crypto_cpp_py==1.4.4
cryptography==45.0.5
curl_cffi==0.11.4
cvxopt==1.3.2
cycler==0.12.1
cytoolz==1.0.1
dateparser==1.2.2
dill==0.4.0
dnspython==2.7.0
ecdsa==0.18.0
eod==0.2.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
fonttools==4.58.5
frozendict==2.4.6
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.44
graphviz==0.21
h11==0.16.0
h2==4.2.0
hexbytes==1.3.1
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lark==1.2.2
line_profiler==4.2.0
llvmlite==0.42.0
lxml==6.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-oneofschema==3.0.1
marshmallow_dataclass==8.6.1
matplotlib==3.10.3
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.18
multitasking==0.0.11
mypy_extensions==1.1.0
narwhals==1.46.0
numba==0.59.1
numpy==1.26.4
numpy-ext==0.9.9
numpy-ringbuffer==0.2.2
onecache==0.7.1
orjson==3.10.18
packaging==25.0
pandas==2.2.3
parsimonious==0.10.0
pathos==0.3.4
patsy==1.0.1
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
poseidon_py==0.1.4
pox==0.3.6
ppft==1.7.7
propcache==0.3.2
protobuf==6.31.1
pyarrow==20.0.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pymongo==4.13.2
pyparsing==3.2.3
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyunormalize==16.0.0
pywin32==306
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rlp==4.1.0
rpds-py==0.26.0
s3transfer==0.10.4
scikit-learn==1.7.0
scipy==1.13.1
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
starknet-crypto-py==0.2.0
starknet-py==0.22.0
statsmodels==0.14.5
streamlit==1.46.1
sympy==1.11.1
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
toolz==1.0.0
tornado==6.5.1
tqdm==4.67.1
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
urllib3==2.5.0
v20==3.0.25.0
watchdog==6.0.0
web3==7.0.0
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.2
yarl==1.20.1
yfinance==0.2.65
